mcp-servers/cross-evaluator/pyproject.toml
# uv project config for cross-evaluator MCP server
[project]
name = "cross-evaluator"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0",
    "google-generativeai>=0.8",
    "openai>=1.50",
    "tiktoken>=0.7",
    "pydantic>=2.0",
]

[project.scripts]
cross-evaluator = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
